feat: ngify

+++ CommonLibF4/include/F4SE/Impl/PCH.h
@@ -22,8 +22,10 @@
 #include <format>
 #include <fstream>
 #include <functional>
-#include <initializer_list>
 #include <intrin.h>
+#include <iomanip>
+#include <ios>
+#include <istream>
 #include <iterator>
 #include <limits>
 #include <locale>
@@ -31,11 +33,11 @@
 #include <memory>
 #include <mutex>
 #include <new>
-#include <numbers>
 #include <numeric>
 #include <optional>
 #include <random>
-#include <ranges>
+#include <regex>
+#include <set>
 #include <source_location>
 #include <span>
 #include <sstream>
@@ -56,7 +58,7 @@ static_assert(
 	std::is_integral_v<std::time_t> && sizeof(std::time_t) == sizeof(std::size_t),
 	"wrap std::time_t instead");
 
-#pragma warning(push, 0)
+#pragma warning(push)
 #include <mmio/mmio.hpp>
 #include <spdlog/spdlog.h>
 #pragma warning(pop)
@@ -112,13 +114,14 @@ namespace F4SE
 			requires(
 				const K&                    a_transparent,
 				const typename C::key_type& a_key,
-				typename C::key_compare&    a_compare) {
-				typename C::key_compare::is_transparent;
-				// clang-format off
+				typename C::key_compare&    a_compare)
+		// clang-format off
+		{
+			typename C::key_compare::is_transparent;
 			{ a_compare(a_transparent, a_key) } -> std::convertible_to<bool>;
 			{ a_compare(a_key, a_transparent) } -> std::convertible_to<bool>;
-				// clang-format on
-			};
+		};
+		// clang-format on
 
 		namespace nttp
 		{
@@ -155,6 +158,7 @@ namespace F4SE
 
 				[[nodiscard]] consteval const_reference back() const noexcept { return (*this)[size() - 1]; }
 				[[nodiscard]] consteval const_pointer   data() const noexcept { return c; }
+				[[nodiscard]] consteval bool            empty() const noexcept { return this->size() == 0; }
 				[[nodiscard]] consteval const_reference front() const noexcept { return (*this)[0]; }
 				[[nodiscard]] consteval size_type       length() const noexcept { return N; }
 				[[nodiscard]] consteval size_type       size() const noexcept { return length(); }
@@ -310,16 +314,132 @@ namespace F4SE
 
 		template <class... Args>
 		enumeration(Args...) -> enumeration<
-			std::common_type_t<Args...>,
-			std::underlying_type_t<
-				std::common_type_t<Args...>>>;
+								 std::common_type_t<Args...>,
+								 std::underlying_type_t<
+									 std::common_type_t<Args...>>>;
+
 	}
 }
+
+#define F4SE_MAKE_LOGICAL_OP(a_op, a_result)                                                                    \
+	template <class E, class U1, class U2>                                                                      \
+	[[nodiscard]] constexpr a_result operator a_op(enumeration<E, U1> a_lhs, enumeration<E, U2> a_rhs) noexcept \
+	{                                                                                                           \
+		return a_lhs.get() a_op a_rhs.get();                                                                    \
+	}                                                                                                           \
+                                                                                                                \
+	template <class E, class U>                                                                                 \
+	[[nodiscard]] constexpr a_result operator a_op(enumeration<E, U> a_lhs, E a_rhs) noexcept                   \
+	{                                                                                                           \
+		return a_lhs.get() a_op a_rhs;                                                                          \
+	}
+
+#define F4SE_MAKE_ARITHMETIC_OP(a_op)                                                        \
+	template <class E, class U>                                                              \
+	[[nodiscard]] constexpr auto operator a_op(enumeration<E, U> a_enum, U a_shift) noexcept \
+		-> enumeration<E, U>                                                                 \
+	{                                                                                        \
+		return static_cast<E>(static_cast<U>(a_enum.get()) a_op a_shift);                    \
+	}                                                                                        \
+                                                                                             \
+	template <class E, class U>                                                              \
+	constexpr auto operator a_op##=(enumeration<E, U>& a_enum, U a_shift) noexcept           \
+		-> enumeration<E, U>&                                                                \
+	{                                                                                        \
+		return a_enum = a_enum a_op a_shift;                                                 \
+	}
+
+#define F4SE_MAKE_ENUMERATION_OP(a_op)                                                                      \
+	template <class E, class U1, class U2>                                                                  \
+	[[nodiscard]] constexpr auto operator a_op(enumeration<E, U1> a_lhs, enumeration<E, U2> a_rhs) noexcept \
+		-> enumeration<E, std::common_type_t<U1, U2>>                                                       \
+	{                                                                                                       \
+		return static_cast<E>(static_cast<U1>(a_lhs.get()) a_op static_cast<U2>(a_rhs.get()));              \
+	}                                                                                                       \
+                                                                                                            \
+	template <class E, class U>                                                                             \
+	[[nodiscard]] constexpr auto operator a_op(enumeration<E, U> a_lhs, E a_rhs) noexcept                   \
+		-> enumeration<E, U>                                                                                \
+	{                                                                                                       \
+		return static_cast<E>(static_cast<U>(a_lhs.get()) a_op static_cast<U>(a_rhs));                      \
+	}                                                                                                       \
+                                                                                                            \
+	template <class E, class U>                                                                             \
+	[[nodiscard]] constexpr auto operator a_op(E a_lhs, enumeration<E, U> a_rhs) noexcept                   \
+		-> enumeration<E, U>                                                                                \
+	{                                                                                                       \
+		return static_cast<E>(static_cast<U>(a_lhs) a_op static_cast<U>(a_rhs.get()));                      \
+	}                                                                                                       \
+                                                                                                            \
+	template <class E, class U1, class U2>                                                                  \
+	constexpr auto operator a_op##=(enumeration<E, U1>& a_lhs, enumeration<E, U2> a_rhs) noexcept           \
+		-> enumeration<E, U1>&                                                                              \
+	{                                                                                                       \
+		return a_lhs = a_lhs a_op a_rhs;                                                                    \
+	}                                                                                                       \
+                                                                                                            \
+	template <class E, class U>                                                                             \
+	constexpr auto operator a_op##=(enumeration<E, U>& a_lhs, E a_rhs) noexcept                             \
+		-> enumeration<E, U>&                                                                               \
+	{                                                                                                       \
+		return a_lhs = a_lhs a_op a_rhs;                                                                    \
+	}                                                                                                       \
+                                                                                                            \
+	template <class E, class U>                                                                             \
+	constexpr auto operator a_op##=(E& a_lhs, enumeration<E, U> a_rhs) noexcept                             \
+		-> E&                                                                                               \
+	{                                                                                                       \
+		return a_lhs = *(a_lhs a_op a_rhs);                                                                 \
+	}
+
+#define F4SE_MAKE_INCREMENTER_OP(a_op)                                                       \
+	template <class E, class U>                                                              \
+	constexpr auto operator a_op##a_op(enumeration<E, U>& a_lhs) noexcept                    \
+		-> enumeration<E, U>&                                                                \
+	{                                                                                        \
+		return a_lhs a_op## = static_cast<E>(1);                                             \
+	}                                                                                        \
+                                                                                             \
+	template <class E, class U>                                                              \
+	[[nodiscard]] constexpr auto operator a_op##a_op(enumeration<E, U>& a_lhs, int) noexcept \
+		-> enumeration<E, U>                                                                 \
+	{                                                                                        \
+		const auto tmp = a_lhs;                                                              \
+		a_op##a_op a_lhs;                                                                    \
+		return tmp;                                                                          \
+	}
 
 namespace F4SE
 {
 	namespace stl
 	{
+		template <
+			class E,
+			class U>
+		[[nodiscard]] constexpr auto operator~(enumeration<E, U> a_enum) noexcept
+			-> enumeration<E, U>
+		{
+			return static_cast<E>(~static_cast<U>(a_enum.get()));
+		}
+
+		F4SE_MAKE_LOGICAL_OP(==, bool);
+		F4SE_MAKE_LOGICAL_OP(<=>, std::strong_ordering);
+
+		F4SE_MAKE_ARITHMETIC_OP(<<);
+		F4SE_MAKE_ENUMERATION_OP(<<);
+		F4SE_MAKE_ARITHMETIC_OP(>>);
+		F4SE_MAKE_ENUMERATION_OP(>>);
+
+		F4SE_MAKE_ENUMERATION_OP(|);
+		F4SE_MAKE_ENUMERATION_OP(&);
+		F4SE_MAKE_ENUMERATION_OP(^);
+
+		F4SE_MAKE_ENUMERATION_OP(+);
+		F4SE_MAKE_ENUMERATION_OP(-);
+
+		F4SE_MAKE_INCREMENTER_OP(+);  // ++
+		F4SE_MAKE_INCREMENTER_OP(-);  // --
+
 		template <class T>
 		class atomic_ref :
 			public std::atomic_ref<T>
@@ -429,8 +549,8 @@ namespace F4SE
 			-> std::optional<std::wstring>
 		{
 			const auto cvt = [&](wchar_t* a_dst, std::size_t a_length) {
-				return WinAPI::MultiByteToWideChar(
-					WinAPI::CP_UTF8,
+				return REX::W32::MultiByteToWideChar(
+					REX::W32::CP_UTF8,
 					0,
 					a_in.data(),
 					static_cast<int>(a_in.length()),
@@ -455,8 +575,8 @@ namespace F4SE
 			-> std::optional<std::string>
 		{
 			const auto cvt = [&](char* a_dst, std::size_t a_length) {
-				return WinAPI::WideCharToMultiByte(
-					WinAPI::CP_UTF8,
+				return REX::W32::WideCharToMultiByte(
+					REX::W32::CP_UTF8,
 					0,
 					a_in.data(),
 					static_cast<int>(a_in.length()),
@@ -478,7 +598,6 @@ namespace F4SE
 
 			return out;
 		}
-
 #ifndef __clang__
 		using source_location = std::source_location;
 #else
@@ -537,41 +656,36 @@ namespace F4SE
 			const char*    _function = "";
 		};
 #endif
-
 		inline bool report_and_error(std::string_view a_msg, bool a_fail = true,
-			F4SE::stl::source_location a_loc = F4SE::stl::source_location::current())
+			std::source_location a_loc = std::source_location::current())
 		{
 			const auto body = [&]() -> std::wstring {
 				const std::filesystem::path p = a_loc.file_name();
-				const auto                  filename = p.generic_string();
-				std::string_view            fileview = filename;
+				auto                        filename = p.lexically_normal().generic_string();
 
-				constexpr auto npos = std::string::npos;
-				std::size_t    pos = npos;
-				std::size_t    off = 0;
-				for (const auto& dir : directories) {
-					pos = fileview.find(dir);
-					if (pos != npos) {
-						off = dir.length();
-						break;
-					}
+				const std::regex r{ R"((?:^|[\\\/])(?:include|src)[\\\/](.*)$)" };
+				std::smatch      matches;
+				if (std::regex_search(filename, matches, r)) {
+					filename = matches[1].str();
 				}
 
-				if (pos != npos) {
-					fileview = fileview.substr(pos + off);
-				}
-
-				return std::format("{}({}): {}", fileview, a_loc.line(), a_msg);
+				return utf8_to_utf16(
+					std::format(
+						"{}({}): {}"sv,
+						filename,
+						a_loc.line(),
+						a_msg))
+				    .value_or(L"<character encoding error>"s);
 			}();
 
-			const auto caption = []() -> std::string {
-				std::vector<char> buf;
+			const auto caption = []() {
+				std::vector<wchar_t> buf;
 				buf.reserve(REX::W32::MAX_PATH);
 				buf.resize(REX::W32::MAX_PATH / 2);
 				std::uint32_t result = 0;
 				do {
 					buf.resize(buf.size() * 2);
-					result = REX::W32::GetModuleFileNameA(
+					result = REX::W32::GetModuleFileNameW(
 						REX::W32::GetCurrentModule(),
 						buf.data(),
 						static_cast<std::uint32_t>(buf.size()));
@@ -592,8 +706,18 @@ namespace F4SE
 					a_loc.function_name() },
 				spdlog::level::critical,
 				a_msg);
-			REX::W32::MessageBoxA(nullptr, body.c_str(), (caption.empty() ? nullptr : caption.c_str()), 0);
-			REX::W32::TerminateProcess(REX::W32::GetCurrentProcess(), EXIT_FAILURE);
+
+			if (a_fail) {
+				REX::W32::MessageBoxW(nullptr, body.c_str(), (caption.empty() ? nullptr : caption.c_str()), 0);
+				REX::W32::TerminateProcess(REX::W32::GetCurrentProcess(), EXIT_FAILURE);
+			}
+			return true;
+		}
+
+		[[noreturn]] inline void report_and_fail(std::string_view a_msg,
+			std::source_location                                  a_loc = std::source_location::current())
+		{
+			report_and_error(a_msg, true, a_loc);
 		}
 
 		template <class To, class From>
@@ -634,62 +758,13 @@ namespace F4SE
 				return to;
 			}
 		}
-
-		[[nodiscard]] inline auto utf8_to_utf16(std::string_view a_in) noexcept
-			-> std::optional<std::wstring>
-		{
-			const auto cvt = [&](wchar_t* a_dst, std::size_t a_length) {
-				return REX::W32::MultiByteToWideChar(
-					REX::W32::CP_UTF8,
-					0,
-					a_in.data(),
-					static_cast<int>(a_in.length()),
-					a_dst,
-					static_cast<int>(a_length));
-			};
-
-			const auto len = cvt(nullptr, 0);
-			if (len == 0) {
-				return std::nullopt;
-			}
-
-			std::wstring out(len, '\0');
-			if (cvt(out.data(), out.length()) == 0) {
-				return std::nullopt;
-			}
-
-			return out;
-		}
-
-		[[nodiscard]] inline auto utf16_to_utf8(std::wstring_view a_in) noexcept
-			-> std::optional<std::string>
-		{
-			const auto cvt = [&](char* a_dst, std::size_t a_length) {
-				return REX::W32::WideCharToMultiByte(
-					REX::W32::CP_UTF8,
-					0,
-					a_in.data(),
-					static_cast<int>(a_in.length()),
-					a_dst,
-					static_cast<int>(a_length),
-					nullptr,
-					nullptr);
-			};
-
-			const auto len = cvt(nullptr, 0);
-			if (len == 0) {
-				return std::nullopt;
-			}
-
-			std::string out(len, '\0');
-			if (cvt(out.data(), out.length()) == 0) {
-				return std::nullopt;
-			}
-
-			return out;
-		}
 	}
 }
+
+#undef F4SE_MAKE_INCREMENTER_OP
+#undef F4SE_MAKE_ENUMERATION_OP
+#undef F4SE_MAKE_ARITHMETIC_OP
+#undef F4SE_MAKE_LOGICAL_OP
 
 namespace RE
 {
@@ -702,6 +777,8 @@ namespace REL
 	using namespace std::literals;
 	namespace stl = F4SE::stl;
 }
+
+#define RELOCATION_ID(a_f4, a_ng) REL::RelocationID(a_f4, a_ng)
 
 #include "REL/REL.h"
